# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Atlas Backend API
  version: 0.1.0
  description: >
    OpenAPI documentation for the Atlas backend API. This spec covers the
    auth, users, voice, drafts, analytics, alerts, images, research, and
    trending routes exposed by the Express service.
servers:
  - url: /
    description: Current server
...
security:
  - bearerAuth: []
paths:
...
  /api/alerts/subscriptions:
    get:
      tags:
        - Alerts
      summary: List alert subscriptions
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Alert subscriptions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertSubscriptionsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Alerts
      summary: Create or re-activate an alert subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlertSubscriptionRequest'
      responses:
        '200':
          description: Subscription saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertSubscriptionEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/alerts/subscriptions/{id}:
    patch:
      tags:
        - Alerts
      summary: Update an alert subscription
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlertSubscriptionRequest'
      responses:
        '200':
          description: Subscription updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertSubscriptionEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Alerts
      summary: Delete an alert subscription
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Subscription deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/alerts/feed:
    get:
      tags:
        - Alerts
      summary: Get the recent alert feed
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Alert feed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/alerts/{id}:
    get:
      tags:
        - Alerts
      summary: Get a single alert
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Alert details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
        - Alerts
      summary: Dismiss an alert
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Alert updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Alerts
      summary: Delete an alert
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Alert deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LimitParam:
      name: limit
      in: query
      description: Max number of records to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    OffsetParam:
      name: offset
      in: query
      description: Number of records to skip
      schema:
        type: integer
        minimum: 0
        default: 0
...
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required:
        - ok
        - error
        - timestamp
      properties:
        ok:
          type: boolean
          enum:
            - false
        error:
          type: string
        details:
          nullable: true
          description: Additional validation or provider-specific detail payload.
        timestamp:
          type: string
          format: date-time
    SuccessEnvelope:
      type: object
      required:
        - ok
        - timestamp
      properties:
        ok:
          type: boolean
          enum:
            - true
        timestamp:
          type: string
          format: date-time
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - success
              properties:
                success:
                  type: boolean
...
    AlertSubscriptionType:
      type: string
      enum:
        - CATEGORY
        - ACCOUNT
        - REPORT_TYPE
    AlertDelivery:
      type: string
      enum:
        - PORTAL
        - TELEGRAM
...
    AlertSubscription:
      type: object
      required:
        - id
        - userId
        - type
        - value
        - isActive
        - delivery
        - createdAt
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          $ref: '#/components/schemas/AlertSubscriptionType'
        value:
          type: string
        isActive:
          type: boolean
        delivery:
          type: array
          items:
            $ref: '#/components/schemas/AlertDelivery'
        createdAt:
          type: string
          format: date-time
    Alert:
      type: object
      required:
        - id
        - type
        - title
        - createdAt
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        context:
          type: string
          nullable: true
        draftReply:
          type: string
          nullable: true
        sourceUrl:
          type: string
          nullable: true
        sentiment:
          type: string
          nullable: true
        relevance:
          type: number
          format: float
          nullable: true
        userId:
          type: string
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    CreateAlertSubscriptionRequest:
      type: object
      required:
        - type
        - value
      properties:
        type:
          $ref: '#/components/schemas/AlertSubscriptionType'
        value:
          type: string
          minLength: 1
        delivery:
          type: array
          items:
            $ref: '#/components/schemas/AlertDelivery'
    UpdateAlertSubscriptionRequest:
      type: object
      properties:
        isActive:
          type: boolean
        delivery:
          type: array
          items:
            $ref: '#/components/schemas/AlertDelivery'
...
    AlertSubscriptionsResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - subscriptions
              properties:
                subscriptions:
                  type: array
                  items:
                    $ref: '#/components/schemas/AlertSubscription'
    AlertSubscriptionEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - subscription
              properties:
                subscription:
                  $ref: '#/components/schemas/AlertSubscription'
    AlertsResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - alerts
              properties:
                alerts:
                  type: array
                  items:
                    $ref: '#/components/schemas/Alert'
    AlertEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - alert
              properties:
                alert:
                  $ref: '#/components/schemas/Alert'
...